Clean up `Boolean` too
