+++ src/fields/quadratic_extension.rs
@@ -377,7 +377,7 @@ where
     fn is_eq(&self, other: &Self) -> Result<Boolean<P::BasePrimeField>, SynthesisError> {
         let b0 = self.c0.is_eq(&other.c0)?;
         let b1 = self.c1.is_eq(&other.c1)?;
-        b0.and(&b1)
+        Ok(b0 & b1)
     }
 
     #[inline]
@@ -400,9 +400,7 @@ where
         condition: &Boolean<P::BasePrimeField>,
     ) -> Result<(), SynthesisError> {
         let is_equal = self.is_eq(other)?;
-        is_equal
-            .and(condition)?
-            .enforce_equal(&Boolean::Constant(false))
+        (is_equal & condition).enforce_equal(&Boolean::FALSE)
     }
 }
