+++ src/groups/curves/short_weierstrass/mod.rs
@@ -182,7 +182,7 @@ where
             //                 `z_inv * self.z = 0` if `self.is_zero()`.
             //
             // Thus, `z_inv * self.z = !self.is_zero()`.
-            z_inv.mul_equals(&self.z, &F::from(infinity.not()))?;
+            z_inv.mul_equals(&self.z, &F::from(!&infinity))?;
 
             let non_zero_x = &self.x * &z_inv;
             let non_zero_y = &self.y * &z_inv;
@@ -735,9 +735,9 @@ where
     ) -> Result<Boolean<<P::BaseField as Field>::BasePrimeField>, SynthesisError> {
         let x_equal = (&self.x * &other.z).is_eq(&(&other.x * &self.z))?;
         let y_equal = (&self.y * &other.z).is_eq(&(&other.y * &self.z))?;
-        let coordinates_equal = x_equal.and(&y_equal)?;
-        let both_are_zero = self.is_zero()?.and(&other.is_zero()?)?;
-        both_are_zero.or(&coordinates_equal)
+        let coordinates_equal = x_equal & y_equal;
+        let both_are_zero = self.is_zero()? & other.is_zero()?;
+        Ok(both_are_zero | coordinates_equal)
     }
 
     #[inline]
@@ -749,12 +749,9 @@ where
     ) -> Result<(), SynthesisError> {
         let x_equal = (&self.x * &other.z).is_eq(&(&other.x * &self.z))?;
         let y_equal = (&self.y * &other.z).is_eq(&(&other.y * &self.z))?;
-        let coordinates_equal = x_equal.and(&y_equal)?;
-        let both_are_zero = self.is_zero()?.and(&other.is_zero()?)?;
-        both_are_zero
-            .or(&coordinates_equal)?
-            .conditional_enforce_equal(&Boolean::Constant(true), condition)?;
-        Ok(())
+        let coordinates_equal = x_equal & y_equal;
+        let both_are_zero = self.is_zero()? & other.is_zero()?;
+        (both_are_zero | coordinates_equal).conditional_enforce_equal(&Boolean::TRUE, condition)
     }
 
     #[inline]
@@ -765,9 +762,7 @@ where
         condition: &Boolean<<P::BaseField as Field>::BasePrimeField>,
     ) -> Result<(), SynthesisError> {
         let is_equal = self.is_eq(other)?;
-        is_equal
-            .and(condition)?
-            .enforce_equal(&Boolean::Constant(false))
+        (is_equal & condition).enforce_equal(&Boolean::FALSE)
     }
 }
